refactor: Inserter
