+++ src/Parser.ts
@@ -1,7 +1,7 @@
 import * as fs from 'fs';
+import Inserter from './Inserter';
 
 export default class Parser {
-    readonly VERSION_SYMBOLS = [ "<", "<=", "=", "!=", ">", ">=", "~>" ];
     content: {[key: string]: Array<{[key: string]: string}>} = {};
     root: string = "dependencies";
     
@@ -9,50 +9,13 @@ export default class Parser {
         return line.startsWith("gem ");
     }
 
-    includeName(dependency: {[key: string]: string}, lineArray: string[]): void {
-        dependency["name"] = lineArray[0].replace("gem", "").replaceAll('"', "").trim();
-    }
-
-    includeVersion(dependency: {[key: string]: string}, lineArray: string[]): void {
-        let versions: string[] = [];
-        let versionsIndexes: number[] = [];
-
-        for (let [index, elem] of lineArray.entries()) {
-            let hasDigit = elem.match(/\d/);
-
-            if (!elem.includes(" ") && hasDigit && hasDigit["index"]) {
-                elem = elem.slice(0, hasDigit["index"]) + " " + elem.slice(hasDigit["index"]);
-            }
-
-            for (let symbol of this.VERSION_SYMBOLS) {
-                if (elem.includes("\"" + symbol + " ")) {
-                    versions.push(elem.replaceAll('"', ''));
-                    versionsIndexes.push(index);
-                }
-            }
-        }
-
-        dependency["version"] = versions.join(", ");
-    }
-
     parseDependency(line: string): {[key: string]: string} {
         let lineArray: string[] = line.split(",");
         let dependency: {[key: string]: string} = {};
-        let version_detected: boolean = false;
-
-        for (let symbol of this.VERSION_SYMBOLS) {
-            if (line.includes(symbol)) {
-                version_detected = true;
-            }
-        }
     
         lineArray = lineArray.map(elem => elem.trim());
 
-        this.includeName(dependency, lineArray);
-
-        if (version_detected) {
-            this.includeVersion(dependency, lineArray);
-        }
+        new Inserter(dependency, lineArray).name().version();
 
         return dependency;
     }
